Fix oss-fuzz build failure
A warning of a uint64_t -> size_t is causing the build to fail on 32-bit builds

Bug: oss-fuzz:45451
Reviewed-on: https://dawn-review.googlesource.com/c/tint/+/83841
Kokoro: Kokoro <noreply+kokoro@google.com>
Commit-Queue: Ben Clayton <bclayton@google.com>

+++ src/tint/sem/bool_type.cc
@@ -28,7 +28,7 @@ Bool::Bool(Bool&&) = default;
 Bool::~Bool() = default;
 
 size_t Bool::Hash() const {
-  return TypeInfo::Of<Bool>().full_hashcode;
+  return static_cast<size_t>(TypeInfo::Of<Bool>().full_hashcode);
 }
 
 bool Bool::Equals(const Type& other) const {
