+++ src/tint/sem/f32_type.cc
@@ -28,7 +28,7 @@ F32::F32(F32&&) = default;
 F32::~F32() = default;
 
 size_t F32::Hash() const {
-  return TypeInfo::Of<F32>().full_hashcode;
+  return static_cast<size_t>(TypeInfo::Of<F32>().full_hashcode);
 }
 
 bool F32::Equals(const Type& other) const {
