+++ src/tint/sem/i32_type.cc
@@ -28,7 +28,7 @@ I32::I32(I32&&) = default;
 I32::~I32() = default;
 
 size_t I32::Hash() const {
-  return TypeInfo::Of<I32>().full_hashcode;
+  return static_cast<size_t>(TypeInfo::Of<I32>().full_hashcode);
 }
 
 bool I32::Equals(const Type& other) const {
